+++ src/ebird/pages/__init__.py
@@ -7,3 +7,4 @@ __version__ = '0.1.0'
 # Import all the functions that make up the public API.
 # noinspection PyUnresolvedReferences
 from ebird.pages.checklists import get_checklist
+from ebird.pages.recent import get_recent_checklists